14/07 14:15

+++ src/main/java/dao/UserDao.java
@@ -11,7 +11,7 @@ public interface UserDao {
 
     void edit(User user) throws SQLException;
 
-    boolean delete(int id) throws SQLException;
+    boolean delete(long id) throws SQLException;
 
     User get(String username);
